Created methods to access data from the database

+++ DataAccessLibrary/Models/DebtIncreaseModel.cs
@@ -6,6 +6,6 @@ namespace DataAccessLibrary.Models
 {
     class DebtIncreaseModel:DebtModel
     {
-        public int Increase { get; set; }
+        public long Increase { get; set; }
     }
 }
